refactor: move files service back to files plugin
