--- imports/plugins/core/core/server/no-meteor/services/files/index.js
@@ -1,7 +1,0 @@
-import ReactionService from "/imports/node-app/core/ReactionService";
-import startup from "./startup";
-
-const service = new ReactionService({ startup });
-service.start();
-
-export default service;
