+++ imports/plugins/core/files/server/fileCollections.js
@@ -12,7 +12,7 @@ import {
 } from "@reactioncommerce/file-collections";
 import { AbsoluteUrlMixin } from "/imports/plugins/core/core/server/Reaction/absoluteUrl";
 import { MediaRecords } from "/lib/collections";
-import setUpFileCollections from "/imports/plugins/core/core/server/no-meteor/services/files/setUpFileCollections";
+import setUpFileCollections from "./no-meteor/setUpFileCollections";
 
 /**
  * Functions and objects related to file upload, download, and storage
