+++ imports/plugins/core/files/server/no-meteor/setUpFileCollections.js
